update: 自分以外のつぶやきは編集・削除できない機能完成

+++ app/Http/Controllers/Tweet/IndexController.php
@@ -3,16 +3,18 @@
 namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
+use App\Services\TweetService;
 use App\Http\Requests\Tweet\CreateRequest;
 use App\Http\Requests\Tweet\UpdateRequest;
 use App\Models\Tweet;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class IndexController extends Controller
 {
-    public function index(Request $request)
+    public function index(Request $request, TweetService $tweetService)
     {
-        $tweets = Tweet::orderBy('created_at', 'DESC')->get();
+        $tweets = $tweetService->getTweets();
 
         return view('tweet.index', ['tweets' => $tweets]);
     }
@@ -35,16 +37,26 @@ class IndexController extends Controller
         return redirect()->route('tweet.index');
     }
 
-    public function update(Request $request)
+    public function update(Request $request, TweetService $tweetService)
     {
         $tweetId = (int)$request->route('tweetId');
+
+        if(!$tweetService->checkOwnTweet($request->user()->id, $tweetId)){
+            throw new AccessDeniedHttpException();
+        }
+
         $tweet   = Tweet::where('tweet_id', $tweetId)->firstOrFail();
 
         return view('tweet.update', ['tweet' => $tweet]);
     }
 
-    public function put(UpdateRequest $request)
+    public function put(UpdateRequest $request, TweetService $tweetService)
     {
+
+        if(!$tweetService->checkOwnTweet($request->user()->id, $request->id())){
+            throw new AccessDeniedHttpException();
+        }
+
         $tweet          = Tweet::where('tweet_id', $request->id())->firstOrFail();
         $tweet->content = $request->tweet();
         $tweet->save();
@@ -54,9 +66,14 @@ class IndexController extends Controller
              ->with('feedback.success', 'つぶやきを編集しました');
     }
 
-    public function delete(Request $request)
+    public function delete(Request $request, TweetService $tweetService)
     {
         $tweetId = (int)$request->route('tweetId');
+
+        if(!$tweetService->checkOwnTweet($request->user()->id, $tweetId)){
+            throw new AccessDeniedHttpException();
+        }
+
         $tweet   = Tweet::where('tweet_id', $tweetId)->firstOrFail();
         $tweet->delete();
